+++ blf_lib/src/blf/versions/destiny.rs
@@ -1,0 +1,3 @@
+pub mod v36735_13_12_02_1953_alpha;
+
+pub const k_title_destiny: &str = "Destiny";